修复报表外部传入带小数的参数值时精度计算有误的BUG

+++ ureport2-core/src/main/java/com/bstek/ureport/definition/datasource/DataType.java
@@ -49,7 +49,7 @@ public enum DataType {
 			if(obj instanceof Float){
 				return (Float)obj;
 			}else{
-				return Utils.toBigDecimal(obj).floatValue();
+				return Utils.toBigDecimal(obj).doubleValue();
 			}
 		case Integer:
 			if(obj.toString().equals("")){
